Dijelovi frontenda odvojeni u komponente i custom hookove
